Refactor: 프롬프트, 파라미터 관련 매개변수 통합
